+++ neoforge/src/main/java/net/mehvahdjukaar/moonlight/api/client/neoforge/ForgeFluidTypeHelper.java
@@ -34,14 +34,8 @@ public class ForgeFluidTypeHelper {
         }
     }
 
-
-    /**
-     * Default constructor.
-     *
-     * @param properties the general properties of the fluid type
-     */
     private static FluidType create(ModFlowingFluid.Properties properties) {
-      return   new FluidType(FluidType.Properties.create()
+        return new FluidType(FluidType.Properties.create()
                 .adjacentPathType(properties.adjacentPathType)
                 .canExtinguish(properties.canExtinguish)
                 .fallDistanceModifier(properties.fallDistanceModifier)